+++ src/main/java/com/bitmart/api/CloudContext.java
@@ -10,7 +10,6 @@ import java.util.Map;
 @Data
 @ToString
 public final class CloudContext {
-    private boolean printLog = false; // Just For Debug
     private boolean isDebug = false;
     private String cloudUrl;
     private long connectTimeoutMilliSeconds = 2000;