+++ src/test/java/com/bitmart/data/TestData.java
@@ -1,5 +1,7 @@
-package com.bitmart.api;
+package com.bitmart.data;
 
+import com.bitmart.api.Call;
+import com.bitmart.api.CloudContext;
 import com.bitmart.api.key.CloudKey;
 
 import java.util.HashMap;
@@ -17,11 +19,10 @@ public class TestData {
     public static String API_KEY = "YOUR ACCESS KEY";
     public static String API_SECRET = "YOUR SECRET KEY";
     public static String API_MEMO = "YOUR MEMO";
-    static Call call;
+    public static Call call;
 
-    TestData() {
+    public TestData() {
         CloudContext cloudContext = new CloudContext(CLOUD_URL, new CloudKey(API_KEY, API_SECRET, API_MEMO));
-        cloudContext.setPrintLog(true);// Just for debug
         cloudContext.setDebug(true);// Can use in production, just for debug
         cloudContext.setReadTimeoutMilliSeconds(10000); // Set read timeout
         // Set your custom headers