+++ src/test/java/com/bitmart/websocket/TestFuturesWebSocket.java
@@ -1,22 +1,22 @@
-package com.bitmart.api;
+package com.bitmart.websocket;
 
+import com.bitmart.data.TestData;
 import com.bitmart.api.key.CloudKey;
-import com.bitmart.websocket.WebSocketCallBack;
-import com.bitmart.websocket.ContractWebSocket;
 import com.google.common.collect.ImmutableList;
+import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Test;
 
 import static com.bitmart.websocket.contract.ContractWebSocketConstant.*;
 
+@Slf4j
+public final class TestFuturesWebSocket extends TestData {
 
-public final class TestContractWebSocket extends TestData {
-
-    public class ReceiveMessage extends WebSocketCallBack {
+    public static class ReceiveMessage extends WebSocketCallBack {
         @Override
         public void onMessage(String text) {
-            System.out.println("onMessage---------------->");
-            System.out.println(text);
+            log.info("onMessage------>{}", text);
         }
+
     }
 
     @Test
@@ -24,8 +24,6 @@ public final class TestContractWebSocket extends TestData {
         ContractWebSocket webSocketPrivateClient = new ContractWebSocket(CLOUD_CONTRACT_WS_PRIVATE_URL,
                 new CloudKey(API_KEY, API_SECRET, API_MEMO), new ReceiveMessage());
         webSocketPrivateClient.login();
-
-        Thread.sleep(2000L);
 
         webSocketPrivateClient.subscribe(ImmutableList.of(
                 // private channel
@@ -45,9 +43,9 @@ public final class TestContractWebSocket extends TestData {
         webSocketClient.setIsPrint(true);
         webSocketClient.subscribe(ImmutableList.of(
                 // public channel
-                WS_PUBLIC_CONTRACT_TICKER,
-                createChannel(WS_PUBLIC_CONTRACT_DEPTH5, "BTCUSDT"),
-                createChannel(WS_PUBLIC_CONTRACT_KLINE_1M,"BTCUSDT")
+                // "futures/ticker"
+                //createChannel(WS_PUBLIC_CONTRACT_DEPTH5, "BTCUSDT"),
+                //createChannel(WS_PUBLIC_CONTRACT_KLINE_1M,"BTCUSDT")
         ));
 
         Thread.sleep(120 * 1000L);